requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
langchain==0.1.16
chromadb==0.4.24
openai==1.30.1
gitpython==3.1.43
apscheduler==4.2.0
python-dotenv==1.0.1
pydantic-settings==2.2.1
tiktoken==0.7.0